refactor: move from command to new
