+++ test/new/service/shortcuts/command_shortcut_events/copy_command_test.dart
@@ -4,8 +4,8 @@ import 'package:appflowy_editor/appflowy_editor.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../../infra/clipboard_test.dart';
-import '../../new/infra/testable_editor.dart';
+import '../../../../infra/clipboard_test.dart';
+import '../../../infra/testable_editor.dart';
 
 void main() async {
   late MockClipboard mockClipboard;
